Refactoring __init__ for lanl
